[MIG] sale_fixed_discount: Migration to 15.0
