+++ sale_fixed_discount/models/sale_order_line.py
@@ -1,0 +1,60 @@
+# Copyright 2017-20 ForgeFlow S.L.
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from odoo import _, api, fields, models
+from odoo.exceptions import ValidationError
+
+
+class SaleOrderLine(models.Model):
+    _inherit = "sale.order.line"
+
+    discount_fixed = fields.Float(
+        string="Discount (Fixed)",
+        digits="Product Price",
+        help="Fixed amount discount.",
+    )
+
+    @api.onchange("discount")
+    def _onchange_discount_percent(self):
+        # _onchange_discount method already exists in core,
+        # but discount is not in the onchange definition
+        if self.discount:
+            self.discount_fixed = 0.0
+
+    @api.onchange("discount_fixed")
+    def _onchange_discount_fixed(self):
+        if self.discount_fixed:
+            self.discount = 0.0
+
+    @api.constrains("discount", "discount_fixed")
+    def _check_only_one_discount(self):
+        for line in self:
+            if line.discount and line.discount_fixed:
+                raise ValidationError(
+                    _("You can only set one type of discount per line.")
+                )
+
+    @api.depends(
+        "product_uom_qty", "discount", "price_unit", "tax_id", "discount_fixed"
+    )
+    def _compute_amount(self):
+        vals = {}
+        for line in self.filtered(
+            lambda l: l.discount_fixed and l.order_id.state not in ["done", "cancel"]
+        ):
+            price = line.price_unit * (1 - (line.discount or 0.0) / 100.0) - (
+                line.discount_fixed or 0.0
+            )
+            vals[line] = {
+                "price_unit": line.price_unit,
+            }
+            line.update({"price_unit": price})
+        res = super()._compute_amount()
+        for line in vals.keys():
+            line.update(vals[line])
+        return res
+
+    def _prepare_invoice_line(self, **optional_values):
+        res = super()._prepare_invoice_line(**optional_values)
+        res.update({"discount_fixed": self.discount_fixed})
+        return res
